+++ infra/modules/azure_merge_roles/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.1.2 (2026-05-08)
+
+### 🩹 Fixes
+
+- Exclude known deprecated permissions from the merged roles such as `Microsoft.Classic` ([#1718](https://github.com/pagopa/dx/pull/1718))
+
+### ❤️ Thank You
+
+- Christian Calabrese
+
+## 0.1.1 (2026-05-08)
+
+### 🩹 Fixes
+
+- Fixed case where actions and not actions would contain the same permission causing duplicated keys errors. ([#1713](https://github.com/pagopa/dx/pull/1713))
+
+### ❤️ Thank You
+
+- Christian Calabrese
+
 ## 0.1.0 (2026-05-07)
 
 ### 🚀 Features